Conflicts:
	src/it/beije/ananke/controller/FirstController.java
	src/it/beije/ananke/controller/RubricaController.java
	src/it/beije/ananke/repository/ContattoRepository.java

+++ src/it/beije/ananke/model/Order.java
@@ -13,30 +13,30 @@ public class Order {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(name="id")
-	private int id;
+	private Integer id;
 	@Column(name="user_id")
-	private int user_id;
+	private Integer user_id;
 	@Column(name="amount")
-	private int amount;
+	private double amount;
 	@Column(name="state")
 	private String state;
 	
-	public int getId() {
+	public Integer getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 	public int getUser_id() {
 		return user_id;
 	}
-	public void setUser_id(int user_id) {
+	public void setUser_id(Integer user_id) {
 		this.user_id = user_id;
 	}
-	public int getAmount() {
+	public double getAmount() {
 		return amount;
 	}
-	public void setAmount(int amount) {
+	public void setAmount(double amount) {
 		this.amount = amount;
 	}
 	public String getState() {
